Added the NpcAiPack interface, Merchant pack (#90)
Adding "ai packages" for enhanced npc interaction. The merchant now
exists and can be edited in BuildMode, but without Buy and Sell
commands, npcs that are merchants won't be accessable to players.

+++ src/shinymud/lib/shiny_server.py
@@ -14,7 +14,7 @@ format = "%(asctime)s %(levelname)s %(name)s %(funcName)s %(lineno)d| %(message)
 logging.basicConfig(filename=LOG_FILE, level=LOG_LEVEL, format=format)
 logger = logging.getLogger('shiny_server')
 world = World()
-world.db.delete('from game_item where owner=null and container=null')
+world.db.delete('from game_item where (owner is null or owner=\'None\') and container is null')
 
 # load the entities in the world from the database
 for area in world.db.select("* from area"):
